LayoutTests/imported/w3c/web-platform-tests/web-animations/animation-model/animation-types/accumulation-per-property-002-expected.txt


PASS Setup
PASS isolation (type: discrete) has testAccumulation function
PASS isolation: "isolate" onto "auto"
PASS isolation: "auto" onto "isolate"
PASS justify-content (type: discrete) has testAccumulation function
PASS justify-content: "end" onto "start"
PASS justify-content: "start" onto "end"
PASS justify-items (type: discrete) has testAccumulation function
PASS justify-items: "end" onto "start"
PASS justify-items: "start" onto "end"
PASS justify-self (type: discrete) has testAccumulation function
PASS justify-self: "end" onto "start"
PASS justify-self: "start" onto "end"
PASS letter-spacing (type: length) has testAccumulation function
FAIL letter-spacing: length assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
FAIL letter-spacing: length of rem assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
PASS lighting-color (type: color) has testAccumulation function
FAIL lighting-color supports animating as color of rgb() with overflowed  from and to values assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL lighting-color supports animating as color of #RGB assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL lighting-color supports animating as color of hsl() assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL lighting-color supports animating as color of #RGBa assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL lighting-color supports animating as color of rgba() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL lighting-color supports animating as color of hsla() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
PASS list-style-image (type: discrete) has testAccumulation function
PASS list-style-image: "url("http://localhost/test-2")" onto "url("http://localhost/test-1")"
PASS list-style-image: "url("http://localhost/test-1")" onto "url("http://localhost/test-2")"
PASS list-style-position (type: discrete) has testAccumulation function
PASS list-style-position: "outside" onto "inside"
PASS list-style-position: "inside" onto "outside"
PASS list-style-type (type: discrete) has testAccumulation function
PASS list-style-type: "square" onto "circle"
PASS list-style-type: "circle" onto "square"
PASS mix-blend-mode (type: discrete) has testAccumulation function
PASS mix-blend-mode: "screen" onto "multiply"
PASS mix-blend-mode: "multiply" onto "screen"
PASS object-fit (type: discrete) has testAccumulation function
PASS object-fit: "contain" onto "fill"
PASS object-fit: "fill" onto "contain"
PASS offset-distance (type: lengthPercentageOrCalc) has testAccumulation function
FAIL offset-distance: length assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
FAIL offset-distance: length of rem assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
FAIL offset-distance: percentage assert_equals: The value should be 130% at 0ms expected "130%" but got "70%"
FAIL offset-distance: units "%" onto "px" assert_equals: The value should be calc(10% + 10px) at 0ms expected "calc(10% + 10px)" but got "10%"
FAIL offset-distance: units "px" onto "%" assert_equals: The value should be calc(10% + 10px) at 0ms expected "calc(10% + 10px)" but got "10px"
FAIL offset-distance: units "rem" onto "%" assert_equals: The value should be calc(10% + 20px) at 0ms expected "calc(10% + 20px)" but got "20px"
FAIL offset-distance: units "%" onto "rem" assert_equals: The value should be calc(10% + 20px) at 0ms expected "calc(10% + 20px)" but got "10%"
FAIL offset-distance: units "rem" onto "em" assert_equals: The value should be 40px at 0ms expected "40px" but got "20px"
FAIL offset-distance: units "em" onto "rem" assert_equals: The value should be 40px at 0ms expected "40px" but got "20px"
FAIL offset-distance: units "calc" onto "px" assert_equals: The value should be calc(20% + 30px) at 0ms expected "calc(20% + 30px)" but got "calc(20% + 20px)"
FAIL offset-distance: calc assert_equals: The value should be calc(30% + 30px) at 0ms expected "calc(30% + 30px)" but got "calc(20% + 20px)"
PASS order (type: integer) has testAccumulation function
FAIL order: integer assert_equals: The value should be -3 at 0ms expected "-3" but got "-2"
PASS outline-color (type: color) has testAccumulation function
FAIL outline-color supports animating as color of rgb() with overflowed  from and to values assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL outline-color supports animating as color of #RGB assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL outline-color supports animating as color of hsl() assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL outline-color supports animating as color of #RGBa assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL outline-color supports animating as color of rgba() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL outline-color supports animating as color of hsla() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
PASS outline-offset (type: length) has testAccumulation function
FAIL outline-offset: length assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
FAIL outline-offset: length of rem assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
PASS outline-style (type: discrete) has testAccumulation function
PASS outline-style: "dotted" onto "none"
PASS outline-style: "none" onto "dotted"
PASS outline-width (type: length) has testAccumulation function
FAIL outline-width: length assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
FAIL outline-width: length of rem assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
PASS overflow-wrap (type: discrete) has testAccumulation function
PASS overflow-wrap: "break-word" onto "normal"
PASS overflow-wrap: "normal" onto "break-word"
PASS overflow-x (type: discrete) has testAccumulation function
PASS overflow-x: "hidden" onto "visible"
PASS overflow-x: "visible" onto "hidden"
PASS overflow-y (type: discrete) has testAccumulation function
PASS overflow-y: "hidden" onto "visible"
PASS overflow-y: "visible" onto "hidden"
PASS page-break-after (type: discrete) has testAccumulation function
PASS page-break-after: "auto" onto "always"
PASS page-break-after: "always" onto "auto"
PASS page-break-before (type: discrete) has testAccumulation function
PASS page-break-before: "auto" onto "always"
PASS page-break-before: "always" onto "auto"
PASS page-break-inside (type: discrete) has testAccumulation function
PASS page-break-inside: "avoid" onto "auto"
PASS page-break-inside: "auto" onto "avoid"
PASS paint-order (type: discrete) has testAccumulation function
PASS paint-order: "stroke" onto "fill"
PASS paint-order: "fill" onto "stroke"
PASS perspective (type: length) has testAccumulation function
FAIL perspective: length assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
FAIL perspective: length of rem assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
PASS perspective-origin (type: position) has testAccumulation function
FAIL perspective-origin: length pair assert_equals: The value should be 20px 20px at 0ms expected "20px 20px" but got "10px 10px"
FAIL perspective-origin: length pair of rem assert_equals: The value should be 20px 20px at 0ms expected "20px 20px" but got "10px 10px"
FAIL perspective-origin: position of percentage assert_equals: The value should be 1019.1875px 0px at 0ms expected "1019.1875px 0px" but got "548.796875px 0px"
PASS pointer-events (type: discrete) has testAccumulation function
PASS pointer-events: "none" onto "fill"
PASS pointer-events: "fill" onto "none"
PASS position (type: discrete) has testAccumulation function
PASS position: "fixed" onto "absolute"
PASS position: "absolute" onto "fixed"
PASS resize (type: discrete) has testAccumulation function
PASS resize: "horizontal" onto "both"
PASS resize: "both" onto "horizontal"
PASS shape-outside (type: discrete) has testAccumulation function
PASS shape-outside: "url("http://localhost/test-2")" onto "url("http://localhost/test-1")"
PASS shape-outside: "url("http://localhost/test-1")" onto "url("http://localhost/test-2")"
PASS stop-color (type: color) has testAccumulation function
FAIL stop-color supports animating as color of rgb() with overflowed  from and to values assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL stop-color supports animating as color of #RGB assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL stop-color supports animating as color of hsl() assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL stop-color supports animating as color of #RGBa assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL stop-color supports animating as color of rgba() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL stop-color supports animating as color of hsla() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
PASS stop-opacity (type: opacity) has testAccumulation function
FAIL stop-opacity: [0, 1] number assert_equals: The value should be 0.6 at 0ms expected "0.6" but got "0.30000001192092896"
FAIL stop-opacity: [0, 1] number (clamped) assert_equals: The value should be 1 at 0ms expected "1" but got "0.30000001192092896"
PASS stroke-dasharray (type: dasharray) has testAccumulation function
PASS stroke-dasharray: dasharray
PASS stroke-dasharray (type: discrete) has testAccumulation function
PASS stroke-dasharray: "10px, 20px" onto "none"
PASS stroke-dasharray: "none" onto "10px, 20px"
PASS stroke-miterlimit (type: positiveNumber) has testAccumulation function
FAIL stroke-miterlimit: positive number assert_equals: The value should be 2.2 at 0ms expected "2.2" but got "1.100000023841858"
PASS stroke-opacity (type: opacity) has testAccumulation function
FAIL stroke-opacity: [0, 1] number assert_equals: The value should be 0.6 at 0ms expected "0.6" but got "0.30000001192092896"
FAIL stroke-opacity: [0, 1] number (clamped) assert_equals: The value should be 1 at 0ms expected "1" but got "0.30000001192092896"
PASS table-layout (type: discrete) has testAccumulation function
PASS table-layout: "fixed" onto "auto"
PASS table-layout: "auto" onto "fixed"
PASS text-align (type: discrete) has testAccumulation function
PASS text-align: "end" onto "start"
PASS text-align: "start" onto "end"
PASS text-decoration-color (type: color) has testAccumulation function
FAIL text-decoration-color supports animating as color of rgb() with overflowed  from and to values assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL text-decoration-color supports animating as color of #RGB assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL text-decoration-color supports animating as color of hsl() assert_equals: The value should be rgb(255, 128, 128) at 0ms expected "rgb(255, 128, 128)" but got "rgb(255, 0, 0)"
FAIL text-decoration-color supports animating as color of #RGBa assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL text-decoration-color supports animating as color of rgba() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
FAIL text-decoration-color supports animating as color of hsla() assert_equals: The value should be rgb(230, 128, 128) at 0ms expected "rgb(230, 128, 128)" but got "rgba(255, 0, 0, 0.4)"
PASS text-decoration-line (type: discrete) has testAccumulation function
PASS text-decoration-line: "overline" onto "underline"
PASS text-decoration-line: "underline" onto "overline"
PASS text-decoration-style (type: discrete) has testAccumulation function
PASS text-decoration-style: "dotted" onto "solid"
PASS text-decoration-style: "solid" onto "dotted"
PASS text-overflow (type: discrete) has testAccumulation function
PASS text-overflow: "ellipsis" onto "clip"
PASS text-overflow: "clip" onto "ellipsis"
PASS text-shadow (type: textShadowList) has testAccumulation function
FAIL text-shadow: shadow assert_equals: The value should be rgb(240, 240, 240) 20px 20px 20px at 0ms expected "rgb(240, 240, 240) 20px 20px 20px" but got "rgb(120, 120, 120) 10px 10px 10px"
PASS text-transform (type: discrete) has testAccumulation function
PASS text-transform: "uppercase" onto "capitalize"
PASS text-transform: "capitalize" onto "uppercase"
PASS touch-action (type: discrete) has testAccumulation function
PASS touch-action: "none" onto "auto"
PASS touch-action: "auto" onto "none"
PASS transform (type: transformList) has testAccumulation function
FAIL transform: translate assert_approx_equals: expected matrix(1, 0, 0, 1, -100, 0) but got matrix(1, 0, 0, 1, -200, 0): The value should be matrix(1, 0, 0, 1, -100, 0) at 0ms but got matrix(1, 0, 0, 1, -200, 0) expected -100 +/- 0.0001 but got -200
FAIL transform: rotate assert_approx_equals: expected matrix(0.707107, -0.707107, 0.707107, 0.707107, 0, 0) but got matrix(0, -1, 1, 0, 0, 0): The value should be matrix(0.707107, -0.707107, 0.707107, 0.707107, 0, 0) at 0ms but got matrix(0, -1, 1, 0, 0, 0) expected 0.707107 +/- 0.0001 but got 0
FAIL transform: scale assert_approx_equals: expected matrix(-2, 0, 0, -2, 0, 0) but got matrix(-3, 0, 0, -3, 0, 0): The value should be matrix(-2, 0, 0, -2, 0, 0) at 0ms but got matrix(-3, 0, 0, -3, 0, 0) expected -2 +/- 0.0001 but got -3
FAIL transform: skew assert_approx_equals: expected matrix(1, 0.57735, -0.36397, 1, 0, 0) but got matrix(1, 0.36397, -0.57735, 1, 0, 0): The value should be matrix(1, 0.57735, -0.36397, 1, 0, 0) at 0ms but got matrix(1, 0.36397, -0.57735, 1, 0, 0) expected 0.57735 +/- 0.0001 but got 0.36397
FAIL transform: rotate on translate assert_approx_equals: expected matrix(0, 1, -1, 0, 100, 0) but got matrix(0, 1, -1, 0, 0, 0): The value should be matrix(0, 1, -1, 0, 100, 0) at 0ms but got matrix(0, 1, -1, 0, 0, 0) expected 100 +/- 0.0001 but got 0
FAIL transform: translate on rotate assert_approx_equals: expected matrix(0, 1, -1, 0, 100, 0) but got matrix(1, 0, 0, 1, 100, 0): The value should be matrix(0, 1, -1, 0, 100, 0) at 0ms but got matrix(1, 0, 0, 1, 100, 0) expected 0 +/- 0.0001 but got 1
FAIL transform: rotate and translate on rotate assert_approx_equals: expected matrix(0, 1, -1, 0, 0, 0) but got matrix(0.707107, 0.707107, -0.707107, 0.707107, 0, 0): The value should be matrix(0, 1, -1, 0, 0, 0) at 0ms but got matrix(0.707107, 0.707107, -0.707107, 0.707107, 0, 0) expected 0 +/- 0.0001 but got 0.707107
FAIL transform: rotate on rotate and translate assert_approx_equals: expected matrix(0, 1, -1, 0, 0, 100) but got matrix(0.707107, 0.707107, -0.707107, 0.707107, 0, 0): The value should be matrix(0, 1, -1, 0, 0, 100) at 0ms but got matrix(0.707107, 0.707107, -0.707107, 0.707107, 0, 0) expected 0 +/- 0.0001 but got 0.707107
FAIL transform: matrix assert_approx_equals: expected matrix(0, 1, -1, 0, 100, 0) but got matrix(1, 0, 0, 1, 100, 0): The value should be matrix(0, 1, -1, 0, 100, 0) at 0ms but got matrix(1, 0, 0, 1, 100, 0) expected 0 +/- 0.0001 but got 1
FAIL transform: rotate3d assert_approx_equals: expected matrix3d(0.853553, 0.146447, 0.5, 0, 0.146447, 0.853553, -0.5, 0, -0.5, 0.5, 0.707107, 0, 0, 0, 0, 1) but got matrix3d(0.5, 0.5, 0.707107, 0, 0.5, 0.5, -0.707107, 0, -0.707107, 0.707107, 0, 0, 0, 0, 0, 1): The value should be matrix3d(0.853553, 0.146447, 0.5, 0, 0.146447, 0.853553, -0.5, 0, -0.5, 0.5, 0.707107, 0, 0, 0, 0, 1) at 0ms but got matrix3d(0.5, 0.5, 0.707107, 0, 0.5, 0.5, -0.707107, 0, -0.707107, 0.707107, 0, 0, 0, 0, 0, 1) expected 0.853553 +/- 0.0001 but got 0.5
FAIL transform: matrix3d assert_approx_equals: expected matrix3d(0.853553, 0.146447, 0.5, 0, 0.146447, 0.853553, -0.5, 0, -0.5, 0.5, 0.707107, 0, 0, 0, 0, 1) but got matrix3d(0.5, 0.5, 0.707107, 0, 0.5, 0.5, -0.707107, 0, -0.707107, 0.707107, 0, 0, 0, 0, 0, 1): The value should be matrix3d(0.853553, 0.146447, 0.5, 0, 0.146447, 0.853553, -0.5, 0, -0.5, 0.5, 0.707107, 0, 0, 0, 0, 1) at 0ms but got matrix3d(0.5, 0.5, 0.707107, 0, 0.5, 0.5, -0.707107, 0, -0.707107, 0.707107, 0, 0, 0, 0, 0, 1) expected 0.853553 +/- 0.0001 but got 0.5
FAIL transform: none assert_equals: dimension of the matrix: The value should be matrix3d(1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 0, 0, 0, 1, 1) at 0ms but got matrix(1, 0, 0, 1, 0, 0) expected 16 but got 6
PASS transform: non-invertible matrices (non-invertible onto invertible)
PASS transform: non-invertible matrices (invertible onto non-invertible)
PASS transform: non-invertible matrices in matched transform lists (non-invertible onto invertible)
PASS transform: non-invertible matrices in matched transform lists (invertible onto non-invertible)
PASS transform: non-invertible matrices in mismatched transform lists (non-invertible onto invertible)
PASS transform: non-invertible matrices in mismatched transform lists (invertible onto non-invertible)
PASS transform-box (type: discrete) has testAccumulation function
PASS transform-box: "border-box" onto "fill-box"
PASS transform-box: "fill-box" onto "border-box"
PASS transform-style (type: discrete) has testAccumulation function
PASS transform-style: "preserve-3d" onto "flat"
PASS transform-style: "flat" onto "preserve-3d"
PASS rotate (type: rotateList) has testAccumulation function
FAIL rotate without rotation axes assert_equals: The value should be -45deg at 0ms expected "-45deg" but got "-90deg"
FAIL rotate with underlying transform assert_equals: The value should be x 45deg at 0ms expected "x 45deg" but got "x 90deg"
PASS rotate with different rotation axes
PASS translate (type: translateList) has testAccumulation function
FAIL translate assert_equals: The value should be -100px at 0ms expected "-100px" but got "-200px"
FAIL translate with transform assert_equals: The value should be -100px at 0ms expected "-100px" but got "-200px"
PASS scale (type: scaleList) has testAccumulation function
FAIL scale with two unspecified values assert_equals: The value should be -2 at 0ms expected "-2" but got "-3"
FAIL scale with one unspecified value assert_equals: The value should be -2 at 0ms expected "-2" but got "-3"
FAIL scale assert_equals: The value should be 0 -1 -2 at 0ms expected "0 -1 -2" but got "-1 -2 -3"
PASS visibility (type: visibility) has testAccumulation function
PASS visibility: onto "visible"
PASS visibility: onto "hidden"
PASS white-space (type: discrete) has testAccumulation function
PASS white-space: "nowrap" onto "pre"
PASS white-space: "pre" onto "nowrap"
PASS word-break (type: discrete) has testAccumulation function
PASS word-break: "break-all" onto "keep-all"
PASS word-break: "keep-all" onto "break-all"
PASS word-spacing (type: lengthPercentageOrCalc) has testAccumulation function
FAIL word-spacing: length assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
FAIL word-spacing: length of rem assert_equals: The value should be 20px at 0ms expected "20px" but got "10px"
FAIL word-spacing: percentage assert_equals: The value should be 130% at 0ms expected "130%" but got "1.75px"
FAIL word-spacing: units "%" onto "px" assert_equals: The value should be calc(10% + 10px) at 0ms expected "calc(10% + 10px)" but got "0.25px"
FAIL word-spacing: units "px" onto "%" assert_equals: The value should be calc(10% + 10px) at 0ms expected "calc(10% + 10px)" but got "10px"
FAIL word-spacing: units "rem" onto "%" assert_equals: The value should be calc(10% + 20px) at 0ms expected "calc(10% + 20px)" but got "20px"
FAIL word-spacing: units "%" onto "rem" assert_equals: The value should be calc(10% + 20px) at 0ms expected "calc(10% + 20px)" but got "0.25px"
FAIL word-spacing: units "rem" onto "em" assert_equals: The value should be 40px at 0ms expected "40px" but got "20px"
FAIL word-spacing: units "em" onto "rem" assert_equals: The value should be 40px at 0ms expected "40px" but got "20px"
FAIL word-spacing: units "calc" onto "px" assert_equals: The value should be calc(20% + 30px) at 0ms expected "calc(20% + 30px)" but got "10px"
FAIL word-spacing: calc assert_equals: The value should be calc(30% + 30px) at 0ms expected "calc(30% + 30px)" but got "0px"
